Improve sampler allocation.
Added my new idea for caching to the sampler allocator as
a simpler test run for it. Looks pretty good. I will add
this to the higher level texture and vertex allocation (i.e
not the Allocator types themselves, but to the map that uses
the allocator, so it can choose what to keep allocated, as
opposed to the Allocator itself being overworked...)