All seems good in the world.
Tested everything up to staging queue submit and it
all seems to be working...
Upload stuff is untested, but it functions actually
in a simpler way because no file io, which is what has
caused me the last 4 hours of trouble as I assure
myself that everything is functioning...
Upload stuff will be tested once I have test case,
i.e. rendering stuff is setup.

OMG!! This is such a relief to complete. This is
almost a real system!!