Adjust weights bug fix.
I think, in going over sampler code, I found a weight wrapping
bug in the allocation weight adjustment code. I think it was
a copy paste error from copying sampler weight code to the
allocation functions, as the clearing of the top bits does not
seem like something that I would need to do for the allocations.